Create test Utils module where functions can be reused
